requirements.txt
torch==2.3.1
torchvision==0.18.1
--extra-index-url https://download.pytorch.org/whl/cu121
tqdm==4.66.4
numpy==1.24.3
matplotlib==3.7.2
pillow==10.4.0
pycocotools==2.0.7
pandas==2.0.3
requests==2.32.2
accelerate==1.11.0